requirements.txt
python-dotenv
langchain
openai
fpdf